BigDecimal for prices instead of double

+++ src/main/java/com/gmail/scanner/model/ScanResult.java
@@ -1,12 +1,13 @@
 package com.gmail.scanner.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ScanResult(String group,
                          String period,
                          List<SourceResult> sources,
-                         double totalSum,
-                         double avgMonth,
+                         BigDecimal totalSum,
+                         BigDecimal avgMonth,
                          String msg) {
 
 }
